+++ src/opts.c
@@ -10,6 +10,7 @@
 #define RAWSTOR_OPTS_SO_RCVTIMEO 5000
 #define RAWSTOR_OPTS_TCP_USER_TIMEOUT 5000
 #define RAWSTOR_OPTS_WAIT_DEVICE_TIMEOUT 5000
+#define RAWSTOR_OPTS_MIRROR_PROBE_INTERVAL 5000
 
 static struct RawstorOpts _rawstor_opts = {};
 
@@ -69,6 +70,14 @@ int rawstor_opts_initialize(const struct RawstorOpts* opts) {
                   RAWSTOR_OPTS_WAIT_DEVICE_TIMEOUT
               );
 
+    _rawstor_opts.mirror_probe_interval =
+        (opts != NULL && opts->mirror_probe_interval != 0)
+            ? opts->mirror_probe_interval
+            : get_env_uint(
+                  "RAWSTOR_OPTS_MIRROR_PROBE_INTERVAL",
+                  RAWSTOR_OPTS_MIRROR_PROBE_INTERVAL
+              );
+
     return 0;
 }
 
@@ -100,4 +109,8 @@ unsigned int rawstor_opts_tcp_user_timeout(void) {
 
 unsigned int rawstor_opts_wait_device_timeout(void) {
     return _rawstor_opts.wait_device_timeout;
+}
+
+unsigned int rawstor_opts_mirror_probe_interval(void) {
+    return _rawstor_opts.mirror_probe_interval;
 }